+++ CHANGELOG.md
@@ -150,6 +150,8 @@
   [#2253](https://github.com/Kong/kong-operator/pull/2253)
 - HybridGateway auto-generated resource names has been revised.
   [#2566](https://github.com/Kong/kong-operator/pull/2566)
+- Update Gateway API to 1.4.0 and k8s libraries to 1.34.
+  [#2451](https://github.com/Kong/kong-operator/pull/2451)
 
 ### Fixes